add support for app bundle, codesign, .ipa archiving

+++ lib/rubixir/rake/builder.rb
@@ -1,6 +1,6 @@
 module Rubixir
   class Builder
-    def compile(config, platform)
+    def build(config, platform)
       datadir = File.join(File.dirname(__FILE__), '../../../data')
       libstatic = File.join(datadir, 'libmacruby-static.a')
       archs = Dir.glob(File.join(datadir, platform, '*.bc')).map do |path|
@@ -17,7 +17,8 @@ module Rubixir
       cxx = File.join(config.platform_dir(platform), 'Developer/usr/bin/g++')
     
       build_dir = File.join(config.build_dir, platform)
-   
+  
+      # Prepare the list of BridgeSupport files needed. 
       bs_flags = ''
       config.frameworks.each do |framework|
         bs_path = File.join(datadir, 'BridgeSupport', framework + '.bridgesupport')
@@ -26,10 +27,11 @@ module Rubixir
         end
       end
 
-      files_compiled = false
       objs = []
+      objs_build_dir = File.join(build_dir, 'objs')
+      FileUtils.mkdir_p(objs_build_dir)
       config.files.each do |path|
-        obj = File.join(build_dir, "#{path}.o")
+        obj = File.join(objs_build_dir, "#{path}.o")
         should_rebuild = (!File.exist?(obj) or File.mtime(path) > File.mtime(obj))
  
         # Generate or retrieve init function.
@@ -37,7 +39,6 @@ module Rubixir
         objs << [obj, init_func]
 
         next unless should_rebuild   
-        files_compiled = true
  
         arch_objs = []
         archs.each do |arch|
@@ -46,14 +47,14 @@ module Rubixir
           raise "Can't locate kernel file" unless File.exist?(kernel)
  
           # Prepare build_dir. 
-          bc = File.join(build_dir, "#{path}.#{arch}.bc")
+          bc = File.join(objs_build_dir, "#{path}.#{arch}.bc")
           FileUtils.mkdir_p(File.dirname(bc))
 
           # LLVM bitcode.
           sh "/usr/bin/env VM_KERNEL_PATH=\"#{kernel}\" #{ruby} #{bs_flags} --emit-llvm \"#{bc}\" #{init_func} \"#{path}\""
  
           # Assembly.
-          asm = File.join(build_dir, "#{path}.#{arch}.s")
+          asm = File.join(objs_build_dir, "#{path}.#{arch}.s")
           llc_arch = case arch
             when 'i386'; 'x86'
             when 'x86_64'; 'x86-64'
@@ -63,7 +64,7 @@ module Rubixir
           sh "#{llc} \"#{bc}\" -o=\"#{asm}\" -march=#{llc_arch} -relocation-model=pic -disable-fp-elim -jit-enable-eh"
  
           # Object.
-          arch_obj = File.join(build_dir, "#{path}.#{arch}.o")
+          arch_obj = File.join(objs_build_dir, "#{path}.#{arch}.o")
           sh "#{cc} -fexceptions -c -arch #{arch} \"#{asm}\" -o \"#{arch_obj}\""
  
           arch_objs << arch_obj
@@ -73,10 +74,6 @@ module Rubixir
         arch_objs_list = arch_objs.map { |x| "\"#{x}\"" }.join(' ')
         sh "lipo -create #{arch_objs_list} -output \"#{obj}\""
       end
-
-      # No need to rebuild the main executable in case no source files changed as well as the static library.
-      main_exec = File.join(build_dir, "main")
-      return if !files_compiled and File.exist?(main_exec) and File.mtime(main_exec) > File.mtime(File.join(datadir, platform, 'libmacruby-static.a'))
 
       # Generate main file.
       main_txt = <<EOS
@@ -126,21 +123,81 @@ EOS
 EOS
  
       # Compile main file.
-      main = File.join(build_dir, 'main.mm')
-      File.open(main, 'w') { |io| io.write(main_txt) }
-      main_o = File.join(build_dir, 'main.o')
       arch_flags = archs.map { |x| "-arch #{x}" }.join(' ')
-      sh "#{cxx} \"#{main}\" #{arch_flags} -fexceptions -fblocks -isysroot \"#{sdk}\" -miphoneos-version-min=4.3 -fobjc-legacy-dispatch -fobjc-abi-version=2 -c -o \"#{main_o}\""
+      main = File.join(objs_build_dir, 'main.mm')
+      main_o = File.join(objs_build_dir, 'main.o')
+      if !(File.exist?(main) and File.exist?(main_o) and File.read(main) == main_txt)
+        File.open(main, 'w') { |io| io.write(main_txt) }
+        sh "#{cxx} \"#{main}\" #{arch_flags} -fexceptions -fblocks -isysroot \"#{sdk}\" -miphoneos-version-min=4.3 -fobjc-legacy-dispatch -fobjc-abi-version=2 -c -o \"#{main_o}\""
+      end
+
+      # Prepare bundle.
+      bundle_path = File.join(build_dir, config.app_name + '.app')
+      FileUtils.mkdir_p(bundle_path)
 
       # Link executable.
+      main_exec = File.join(bundle_path, config.app_name)
       objs_list = objs.map { |path, _| path }.unshift(main_o).map { |x| "\"#{x}\"" }.join(' ')
       frameworks = config.frameworks.map { |x| "-framework #{x}" }.join(' ')
       framework_stubs_objs = []
       config.frameworks.each do |framework|
         stubs_obj = File.join(datadir, platform, "#{framework}_stubs.o")
-        framework_stubs_objs << stubs_obj if File.exist?(stubs_obj)
+        framework_stubs_objs << "\"#{stubs_obj}\"" if File.exist?(stubs_obj)
       end
-      sh "#{cxx} -o #{main_exec} #{objs_list} #{arch_flags} -isysroot \"#{sdk}\" -L#{File.join(datadir, platform)} -lmacruby-static -lobjc -licucore #{frameworks} #{framework_stubs_objs.join(' ')}" 
+      sh "#{cxx} -o \"#{main_exec}\" #{objs_list} #{arch_flags} -isysroot \"#{sdk}\" -L#{File.join(datadir, platform)} -lmacruby-static -lobjc -licucore #{frameworks} #{framework_stubs_objs.join(' ')}"
+
+      # Create bundle/Info.plist.
+      bundle_info_plist = File.join(bundle_path, 'Info.plist')
+      File.open(bundle_info_plist, 'w') { |io| io.write(config.plist_data) }
+      sh "/usr/bin/plutil -convert binary1 \"#{bundle_info_plist}\""
+
+      # Create bundle/PkgInfo.
+      File.open(File.join(bundle_path, 'PkgInfo'), 'w') { |io| io.write(config.pkginfo_data) }
+    end
+
+    def codesign(config, platform)
+      bundle_path = File.join(config.build_dir, platform, config.app_name + '.app')
+      raise unless File.exist?(bundle_path)
+
+      # Create bundle/ResourceRules.plist.
+      resource_rules_plist = File.join(bundle_path, 'ResourceRules.plist')
+      File.open(resource_rules_plist, 'w') do |io|
+        io.write(<<-PLIST)
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+        <key>rules</key>
+        <dict>
+                <key>.*</key>
+                <true/>
+                <key>Info.plist</key>
+                <dict>
+                        <key>omit</key>
+                        <true/>
+                        <key>weight</key>
+                        <real>10</real>
+                </dict>
+                <key>ResourceRules.plist</key>
+                <dict>
+                        <key>omit</key>
+                        <true/>
+                        <key>weight</key>
+                        <real>100</real>
+                </dict>
+        </dict>
+</dict>
+</plist>
+PLIST
+      end
+
+      # Copy the provisioning profile.
+      File.open(File.join(bundle_path, "embedded.mobileprovision"), 'w') do |io|
+        io.write(File.read(config.provisioning_profile))
+      end
+ 
+      # Do the codesigning.
+      sh "/usr/bin/codesign -f -s \"#{config.codesign_certificate}\" --resource-rules=\"#{resource_rules_plist}\" \"#{bundle_path}\""
     end
   end
 end
